First draft of moving menubar to a separate class. Shortcuts don't work.
