+++ quivilib/gui/main.py
@@ -10,22 +10,22 @@ from pubsub import pub as Publisher
 
 from quivilib import meta
 from quivilib import util
+from quivilib.gui.components.menu_bar import QuiviMenuBar
 from quivilib.gui.components.status_bar import QuiviStatusBar
 from quivilib.gui.debug_cache import DebugCacheDialog
 from quivilib.gui.debug_memory import DebugMemoryDialog
 from quivilib.gui.file_list import FileListPanel
 from quivilib.i18n import _
 from quivilib.interface.canvasadapter import CanvasAdapter
-from quivilib.model import Favorites
 from quivilib.model.canvas import PaintedRegion
-from quivilib.model.command import Command, CommandCategory
-from quivilib.model.commandenum import MenuName, CommandName, FitSettings
+from quivilib.model.command import Command
+from quivilib.model.commandenum import FitSettings
 from quivilib.model.container import Item
 from quivilib.model.container.base import BaseContainer
-from quivilib.model.favorites import FavoriteMenuItem
 from quivilib.model.settings import Settings
 from quivilib.resources import images
 from quivilib.util import error_handler
+
 
 def _handle_error(exception, args, kwargs):
     self = args[0]
@@ -54,7 +54,8 @@ class MainWindow(wx.Frame):
         dt = self.QuiviFileDropTarget(self)
         self.SetDropTarget(dt)
         
-        self.menu_bar = wx.MenuBar()
+        #self.menu_bar = QuiviMenuBar(err_fn = lambda e: self.handle_error(e))
+        self.menu_bar = QuiviMenuBar()
         self.SetMenuBar(self.menu_bar)
         
         self.status_bar = QuiviStatusBar(self)
@@ -81,19 +82,8 @@ class MainWindow(wx.Frame):
         self._last_size = self.get_window_size()
         self._last_pos = self.GetPosition()     # NOTE - This has no effect on Wayland
         self._busy = False
-        #List of (id, name) tuples. Filled on the favorites.changed event,
-        #used in the file list popup menu
-        self.favorites_menu_items: list[FavoriteMenuItem] = []
-        self._favorite_menu_count = 0
-        self.update_menu_item = None
+
         self.accel_table = None
-        #Track as a dictionary
-        self.menus: dict[MenuName, wx.Menu] = {}
-        self.menu_names: dict[MenuName, str] = {}
-        #Used for updating translations dynamically. Pair the actual wx objects and the local definitions.
-        self.all_cmd_pairs: list[tuple[Command, wx.MenuItem]] = []
-        #Set by a background task if there is an update available.
-        self.down_url: str|None = None
 
     def bindings_and_subscriptions(self):
         self.panel.Bind(wx.EVT_PAINT, self.on_panel_paint)
@@ -114,12 +104,8 @@ class MainWindow(wx.Frame):
         Publisher.subscribe(self.on_language_changed, 'language.changed')
         Publisher.subscribe(self.on_canvas_changed, 'canvas.changed')
         Publisher.subscribe(self.on_canvas_cursor_changed, 'canvas.cursor.changed')
-        Publisher.subscribe(self.on_menu_built, 'menu.built')
-        Publisher.subscribe(self.on_menu_labels_changed, 'menu.labels.changed')
         Publisher.subscribe(self.on_shortcuts_changed, 'menu.shortcuts.changed')
         Publisher.subscribe(self.on_cmd_context_menu, 'menu.context_menu')
-        Publisher.subscribe(self.on_favorites_changed, 'favorites.changed')
-        #Publisher.subscribe(self.on_favorite_settings_changed, 'settings.changed.Options.PlaceholderSeparateMenu')
         Publisher.subscribe(self.on_settings_loaded, 'settings.loaded')
         Publisher.subscribe(self.on_open_wallpaper_dialog, 'wallpaper.open_dialog')
         Publisher.subscribe(self.on_open_options_dialog, 'options.open_dialog')
@@ -128,7 +114,6 @@ class MainWindow(wx.Frame):
         Publisher.subscribe(self.on_open_debug_memory_dialog, 'debug.open_memory_dialog')
         Publisher.subscribe(self.on_open_about_dialog, 'about.open_dialog')
         Publisher.subscribe(self.on_open_directory_dialog, 'file_list.open_directory_dialog')
-        Publisher.subscribe(self.on_update_available, 'program.update_available')
         Publisher.subscribe(self.on_bg_color_changed, 'settings.loaded')
         Publisher.subscribe(self.on_bg_color_changed, 'settings.changed.Options.CustomBackground')
         Publisher.subscribe(self.on_bg_color_changed, 'settings.changed.Options.CustomBackgroundColor')
@@ -148,7 +133,7 @@ class MainWindow(wx.Frame):
     def canvas_view(self):
         return CanvasAdapter(self.panel)
     
-    def save(self, settings_lst):
+    def save(self, settings_lst: list[tuple[str, str, str]]):
         perspective = self.aui_mgr.SavePerspective()
         settings_lst.append(('Window', 'Perspective', perspective))
         settings_lst.append(('Window', 'MainWindowX', self._last_pos[0]))
@@ -251,11 +236,11 @@ class MainWindow(wx.Frame):
 
     def on_fit_context_menu(self, event: wx.ContextMenuEvent):
         """Appears on right-clicking the status bar"""
-        self.PopupMenu(self.menus[MenuName.FitCtx])
+        self.menu_bar.open_fit_menu()
         
     def on_cmd_context_menu(self):
         """Appears on executing the bindable 'open context menu' command, e.g. middle/right clicking. """
-        self.PopupMenu(self.menus[MenuName.ImgCtx])
+        self.menu_bar.open_context_menu()
         
     def on_busy(self, *, busy):
         if self._busy == busy:
@@ -282,154 +267,10 @@ class MainWindow(wx.Frame):
     def on_canvas_cursor_changed(self, *, cursor: wx.Cursor):
         self.panel.SetCursor(cursor)
 
-    def on_menu_built(self, *, main_menu: list[MenuName], all_menus: list[CommandCategory], commands: list[Command]):
-        """ Turn the model objects into actual wx menu objects and store them locally.
-        These will be used to populate the menu bar (immediately) and context menus (on demand)
-        :param main_menu: The menus that should appear in the menubar. The associated category object will be modified to include the index.
-        :param all_menus: All CommandCategory objects (derived from the MenuDefinition). Order matters as menus may reference previous menus.
-        :param commands: All command objects
-        """
-        menu_lookup = {x.idx: x for x in all_menus}
-        cmd_lookup = {x.ide: x for x in commands if type(x) is Command}
-        #This function should only be called once. But if it is called multiple times, reset state.
-        self.all_cmd_pairs = []
-        #First, create the wx.Menu objects. This is done for everything. Populate self.menus
-        for item in all_menus:
-            #make_menu will also modify all_cmd_pairs
-            wx_menu = self.make_menu(item, menu_lookup, cmd_lookup)
-            self.menus[item.idx] = wx_menu
-            self.menu_names[item.idx] = item.name
-        
-        #Add the appropriate items to self.menu_bar (use Append). Set indices
-        i = 0
-        for idx in main_menu:
-            menu = self.menus[idx]
-            category = menu_lookup[idx]
-            self.menu_bar.Append(menu, category.name)
-            #Need to manually track the id. Searching by name doesn't work if the name can change (translations)
-            #Just counting up is fine as long as there aren't existing menu items, and there shouldn't be.
-            category.menu_idx = i
-            i += 1
-
-        #Create actual bindings for the commands
-        for command in commands:
-            def event_fn(event, cmd=command):
-                try:
-                    cmd()
-                except Exception as e:
-                    self.handle_error(e)
-            self.Bind(wx.EVT_MENU, event_fn, id=command.ide)
-        #This is the number of pre-defined menu items in favorites; everything past this is a favorite.
-        self._favorite_menu_count = self.menus[MenuName.Favorites].GetMenuItemCount()
-
-    def make_menu(self, menu: CommandCategory, all_menus: dict[MenuName, CommandCategory], cmd_lookup: dict[int, Command]) -> wx.Menu:
-        """ Creates the actual wx.Menu for a given CommandCategory.
-        Still requires references to all data, since this may include submenus.
-        """
-        _menu = wx.Menu()
-        for cmd in menu.commands:
-            if cmd is None:
-                _menu.AppendSeparator()
-            # Submenu
-            elif type(cmd) is MenuName:
-                if cmd not in self.menus:
-                    raise Exception(f"Menu {cmd} referenced before it was created.")
-                submenu = self.menus[cmd]
-                data = all_menus[cmd]
-                _menu.AppendSubMenu(submenu, data.name)
-            # Command
-            elif type(cmd) is CommandName:
-                command = cmd_lookup[cmd]
-                style = wx.ITEM_CHECK if command.checkable else wx.ITEM_NORMAL
-                wx_menuitem = _menu.Append(command.ide, command.name_and_shortcut, command.description, style)
-                #Track for later updates (i.e. translations).
-                self.all_cmd_pairs.append((command, wx_menuitem))
-                #If a cmd is in multiple menus, it will bind multiple times. Is this a problem?
-                if command.update_function:
-                    wx.GetApp().Bind(wx.EVT_UPDATE_UI, command.update_function, id=command.ide)
-        return _menu
-
-    def _reset_favorite_menus(self):
-        """The favorites menus are always updated by wiping them out and re-building from scratch.
-        Call this within a 'freeze' block. """
-        favorites_menu = self.menus[MenuName.Favorites]
-
-        reset_submenus = (self.menus[MenuName.FavoritesSub], self.menus[MenuName.PlaceholderSub], self.menus[MenuName.FavoritesCtx], self.menus[MenuName.PlaceholderCtx])
-        for menu in reset_submenus:
-            while menu.GetMenuItemCount() > 0:
-                item = menu.FindItemByPosition(0)
-                menu.Delete(item)
-        # self._favorite_menu_count is the number of submenus in the favorites menu;
-        #      entries bigger than this are the favorites themselves.
-        while favorites_menu.GetMenuItemCount() > self._favorite_menu_count:
-            item = favorites_menu.FindItemByPosition(self._favorite_menu_count)
-            favorites_menu.Delete(item)
-    def on_favorite_settings_changed(self, settings: Settings):
-        """Updates the various menus that display favorites. Called when settings change or when the favorites change."""
-        #The best way to handle this is likely to alternate between adding favorites directly to the menu and the two sub menus.
-        #Trying to do this is giving me wx free errors.
-        pass
-    def _create_favorites(self, favorites: Favorites):
-        """Resets and populates self.favorites_menu_items"""
-        items = favorites.getitems()
-        self.favorites_menu_items = []
-        i = 0
-        for path_key, fav in items:
-            ide = wx.NewId()
-            def event_fn(event: wx.CommandEvent, favorite=fav):
-                try:
-                    Publisher.sendMessage('favorite.open', favorite=favorite, window=self)
-                except Exception as e:
-                    self.handle_error(e)
-
-            name = fav.displayText()
-            if not name:
-                continue
-
-            self.favorites_menu_items.append(FavoriteMenuItem(ide, name, fav))
-            self.Bind(wx.EVT_MENU, event_fn, id=ide)
-            i += 1
-    def on_favorites_changed(self, *, favorites: Favorites, settings: Settings):
-        favorites_menu = self.menus[MenuName.Favorites]
-        fav_only = self.menus[MenuName.FavoritesSub]
-        fav_ctx = self.menus[MenuName.FavoritesCtx]
-        place_only = self.menus[MenuName.PlaceholderSub]
-        place_ctx = self.menus[MenuName.PlaceholderCtx]
-        self._create_favorites(favorites)
-        self.menu_bar.Freeze()
-        try:
-            self._reset_favorite_menus()
-            #Rebuild
-            if self.favorites_menu_items:
-                favorites_menu.AppendSeparator()
-            for item in self.favorites_menu_items:
-                favorites_menu.Append(item.ide, item.name)
-                if item.fav.is_placeholder():
-                    place_only.Append(item.ide, item.name)
-                    place_ctx.Append(item.ide, item.name)
-                else:
-                    fav_only.Append(item.ide, item.name)
-                    fav_ctx.Append(item.ide, item.name)
-        finally:
-            self.menu_bar.Thaw()
-        pass
-
     def on_shortcuts_changed(self, *, accel_table: wx.AcceleratorTable):
         self.accel_table = accel_table
         self.SetAcceleratorTable(self.accel_table)
 
-    def on_menu_labels_changed(self, *, categories: list[CommandCategory]):
-        #Commands (i.e. wx.MenuItem s) use stored data. The menu_bar requires indices; wx.Menu references will not work.
-        for (cmd, wx_item) in self.all_cmd_pairs:
-            wx_item.SetItemLabel(cmd.name)
-            wx_item.SetHelp(cmd.description)
-        for category in categories:
-            #Need to use the idx stored in the category (when the bar is created)
-            #Finding by name isn't reliable if the name can change.
-            midx = category.menu_idx
-            if midx != -1:
-                self.menu_bar.SetMenuLabel(midx, category.name)
-    
     def on_container_opened(self, *, container: BaseContainer):
         self.SetTitle(f'{container.name} - {meta.APPNAME}')
     
@@ -439,10 +280,6 @@ class MainWindow(wx.Frame):
     def on_language_changed(self):
         self.aui_mgr.GetPane('file_list').Caption(_('Files'))
         self.aui_mgr.Update()
-        if self.update_menu_item:
-            self.update_menu_item.SetItemLabel(_('&Download'))
-            self.update_menu_item.SetHelp(_('Go to the download site'))
-            self.menu_bar.SetMenuLabel(self.menu_bar.GetMenuCount()-1, _('&New version available!'))
     
     def on_open_wallpaper_dialog(self, *, choices: list[str], color: wx.Colour):
         from quivilib.gui.wallpaper import WallpaperDialog
@@ -488,16 +325,9 @@ class MainWindow(wx.Frame):
         dialog = AboutDialog(self)
         dialog.ShowModal()
         dialog.Destroy()
-        
-    def on_update_available(self, *, down_url, check_time, version):
-        self.down_url = down_url
-        menu = self.menus[MenuName.Downloads]
-        menu_idx = self.menu_bar.GetMenuCount()
-        self.menu_bar.Append(menu, self.menu_names[MenuName.Downloads])
-        Publisher.sendMessage('menu.item_added', cmd=MenuName.Downloads, idx=menu_idx)
 
     def on_download_update(self):
-        Publisher.sendMessage('program.open_update_site', url=self.down_url)
+        self.menu_bar.on_download_update()
 
     def on_bg_color_changed(self, *, settings: Settings):
         if settings.get('Options', 'CustomBackground') == '1':
